make macos happy

+++ include/Support/Pipeline.h
@@ -295,7 +295,7 @@ struct PushConstantValue {
   // The bytes of this group.
   llvm::SmallVector<char, 4> Data;
   // The offset of this group from the start of the push constant buffer.
-  size_t OffsetInBytes;
+  uint32_t OffsetInBytes;
 };
 
 // Describes the content of the push constant buffer.
@@ -313,7 +313,7 @@ struct PushConstantBlock {
   void getContent(llvm::SmallVectorImpl<uint8_t> &output) const;
 
   // Returns the size in bytes of the whole push constant once laid out.
-  size_t size() const;
+  uint32_t size() const;
 };
 
 struct SpecializationConstant {
